create routing_state base class
